requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==3.6.4.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==4.0.5
django-cors-headers==4.3.1
django-redis==5.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
kombu==5.3.7
mysqlclient==2.1.0
prompt-toolkit==3.0.43
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.4
sqlparse==0.5.0
typing_extensions==4.11.0
vine==5.1.0
wcwidth==0.2.13
